Fix pub_all
Solved by changing the set of actual keys to a list and appending them
rather than trying to build a tuple
Only publishes to keys that contain anything, if the key is empty or 'nil'
then it is not published to. This is a limitation of Redis' KEYS command
In the event that a key has no value I see little point in looking at it,
basically doing an existance test which can be represented by a key
containing true/false.