Convert the scores to use the same function for scoring and modifying
Causes the same UI to be used for both entering scores and changing
them, while still maintaining the inability to accidentally modify
a score during entry.
Now requires a whole match to be entered in one go when scoring.
Allows for half a match to be entered while scoring in the event of
a crash.